% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{build_kfold}
\alias{build_kfold}
\title{build idx matrix for K-fold validation}
\usage{
build_kfold(k, n, seed = NA)
}
\arguments{
\item{k}{for k fold validation}

\item{n}{the total number of rows for dataset}

\item{seed}{random seed for sampling}
}
\value{
`kfold` a matrix with each index of obs in k groups
}
\description{
This function could be use for groupping idx for K-fold
validation
}
\examples{
data(diabetic_data)
kfold <- build_kfold(5, dim(diabetic_data)[1], seed = 103221)
}
