% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{build_random_forest}
\alias{build_random_forest}
\title{build the random forest model for given dataset}
\usage{
build_random_forest(
  data,
  ycol,
  is_kfold = FALSE,
  cv_num = 10,
  seed = NA,
  is_smote = FALSE,
  thre = NA,
  ntree = 100,
  maxnodes = 50,
  mtry = NA
)
}
\arguments{
\item{data}{dataset for modeling}

\item{ycol}{the colname for target column}

\item{is_kfold}{if TRUE, evaluate the model with k-fold evaluation}

\item{cv_num}{for k = cv_num fold validation}

\item{seed}{use for spliting train/test data}

\item{is_smote}{whether upsampling with smote method for solving imbalance}

\item{thre}{classification threshold, when the model gives probability
higher than this value, assign it to the positive, otherwise to negative. If
it's NA, then use the information from ROC (specificities & sensitivities)
to calculated automatically. If it's NA, the threshold would be calculated
automatically based on ROC}

\item{ntree}{the hyp for random forest}

\item{maxnodes}{the hyp for random forest}

\item{mtry}{the hyp for random forest}
}
\value{
`result` a list with `result$mod` trained model; `result$eval_tab`
the model evaluation result; `result$roc_res` the ROC and AUC information;
`result$best_thre` the best threshold used in modeling
}
\description{
This function could be use for training a random forest model
for given dataset, with k-fold validation (if applicable) and evaluate the
result
}
\examples{
data(diabetic_data)
diabetic_data <- diabetic_data[1:100, c('time_in_hospital', 'readmitted')] |>
                 ycol_mapped01(ycol = "readmitted", positive = "YES")
res <- build_random_forest(diabetic_data,
                         "readmitted", is_kfold = TRUE,
                         cv_num = 2, seed = 103221, is_smote = FALSE,
                         thre = 0.2)
}
