% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{build_RandomForest}
\alias{build_RandomForest}
\title{build the random forest model for given dataset}
\usage{
build_RandomForest(
  data,
  ycol,
  is.kfold = FALSE,
  cv_num = 10,
  seed = NA,
  is.smote = FALSE,
  thre = NA
)
}
\arguments{
\item{data}{dataset for modeling}

\item{ycol}{the colname for target column}

\item{is.kfold}{if TRUE, evaluate the model with k-fold evaluation}

\item{cv_num}{for k = cv_num fold validation}

\item{seed}{use for spliting train/test data}

\item{is.smote}{whether upsampling with SMOTE method for solving imbalance}

\item{thre}{classification threshold, when the model gives probability higher
than this value, assign it to the positive, otherwise to negative}
}
\value{
`result` a list with `result$mod` trained model; `result$eval_tab` the
model evaluation result
}
\description{
This function could be use for training a random forest model
for given dataset, with k-fold validation (if applicable) and evaluate the result
}
\examples{
data(diabetic_data)
diabetic_data <- diabetic_data[1:100, c('time_in_hospital', 'readmitted')] |>
                 ycol_mapped01(ycol = "readmitted", positive = "YES")
res <- build_RandomForest(diabetic_data,
                         "readmitted", is.kfold = TRUE,
                         cv_num = 2, seed = 103221, is.smote = FALSE, thre = 0.2)
}
