% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woe.R
\name{build_woe}
\alias{build_woe}
\title{do WOE transformation for given dataset}
\usage{
build_woe(data, ycol = "readmitted", positive = "YES", is.dropsame = TRUE)
}
\arguments{
\item{data}{dataset, need to be data.frame()}

\item{ycol}{the name for the target column}

\item{positive}{the level in target y which denotes the positive class}

\item{is.dropsame}{if TRUE, drop the constant columns after transformation}
}
\value{
processed dataset (data.frame())
}
\description{
This function would do WOE transformation for given data
}
\examples{
data(diabetic_data)
diabetic_data.woe <- build_woe(diabetic_data[1:100, c("num_procedures", "readmitted")],
                               ycol = "readmitted",
                               positive = "YES", is.dropsame = TRUE)
}
