% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_ids_map}
\alias{data_ids_map}
\title{mapping ids with its meaning for dataset}
\usage{
data_ids_map(data, data_map)
}
\arguments{
\item{data}{the original dataset need to do the replacement. need to be data.frame}

\item{data_map}{the map information, need to be data.frame,
with `column` denotes the variable name, `value` denotes the ids valu,
and `description` shows its meaning}
}
\value{
the mapped dataset
}
\description{
This function could be use for replacing the id value in original
dataset with its meaning
}
\examples{
data(diabetic_data)
data(IDs_map)
data.mapped <- data_ids_map(diabetic_data, IDs_map)
}
