% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{encode_ICD9}
\alias{encode_ICD9}
\title{encoding ICD-9 to primary category}
\usage{
encode_ICD9(x, data_icd, maplist = c("diag_1", "diag_2", "diag_3"))
}
\arguments{
\item{x}{the original dataset need to do the transformation, need to be data.frame}

\item{data_icd}{the mapping dataset with ICD-9 codes and its primary category,
should have columns (`description`, `code_start`, `code_end`). refer to the sample
data `ICD9_map` for more information about data format}

\item{maplist}{the vector contains column names which need to be mapped}
}
\value{
the processed dataset
}
\description{
This function could be use for transforming ICD-9 codes in dataset
to its primary category
}
\examples{
data(diabetic_data)
data(ICD9_map)
diabetic_data <- encode_ICD9(x = diabetic_data[1:100, ], data_icd = ICD9_map,
                             maplist = c('diag_1', 'diag_2', 'diag_3'))
}
