% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{eval_model}
\alias{eval_model}
\title{evaluate the classification model}
\usage{
eval_model(pre, gt)
}
\arguments{
\item{pre}{the prediction of the model}

\item{gt}{label for the true value}
}
\value{
`result` a data.frame with `acc`, `precision`, `recall`, `f1`
}
\description{
This function could be use for do evaluation for a classification model
}
\examples{
pred <- c(0,1,1,1,1,0)
gt <- c(1,1,1,1,0,1)
result <- eval_model(pred, gt)
}
