% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{extreme_imbalance_col}
\alias{extreme_imbalance_col}
\title{explore the extreme imbalance columns}
\usage{
extreme_imbalance_col(x, extreme.prop = 0.95, is.table = TRUE, is.drop = TRUE)
}
\arguments{
\item{x}{dataset, need to be data.frame()}

\item{extreme.prop}{if a column with more than extreme.prop obs concentrated in
one single level, it could say that this column is extreme imbalance}

\item{is.table}{if TRUE, print a table to show the information about the
extreme imbalance column found by the rules above}

\item{is.drop}{if TRUE, drop the extreme imbalance column}
}
\value{
processed dataset (data.frame())
}
\description{
This function would do exploration to find extreme imbalance columns
in given dataset
}
\examples{
data(diabetic_data)
diabetic_data <- extreme_imbalance_col(diabetic_data[1:100, ], extreme.prop = .95,
                                 is.drop = FALSE, is.table = TRUE)
}
