% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_factor_col}
\alias{format_factor_col}
\title{format given column to factor}
\usage{
format_factor_col(x)
}
\arguments{
\item{x}{the original column need to do the transformation}
}
\value{
the processed column
}
\description{
This function could be use for transforming given character/category
variables to factor. If the given one is numeric, process would be skipped
}
\examples{
data(diabetic_data)
diabetic_data$readmitted <- format_factor_col(diabetic_data$readmitted)
}
