% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_factor_dataset}
\alias{format_factor_dataset}
\title{format character/category columns to factor for given dataset}
\usage{
format_factor_dataset(x)
}
\arguments{
\item{x}{the original dataset need to do the transformation}
}
\value{
the processed datasaet
}
\description{
This function could be use for transforming the character/
category variables in given dataset to factor. If the column is numeric,
process would be skipped
}
\examples{
data(diabetic_data)
diabetic_data <- format_factor_dataset(diabetic_data[1:100, ])
}
