% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_woe_plot}
\alias{get_woe_plot}
\title{give histogram plot for given columns}
\usage{
get_woe_plot(bins_woe, xcol, ncol = 2)
}
\arguments{
\item{bins_woe}{the bins created by WOE function from `scorecard`}

\item{xcol}{a vector of target column names}

\item{ncol}{the number of plots to show per row}
}
\description{
This function could be used for giving the stack barplot with
line to explore more about the WOE encoding for specific variable
}
\examples{
data(diabetic_data)
diabetic_data[1:1, "metformin.rosiglitazone"] <- "YES"
res_woe_temp <- build_woe(diabetic_data[1:100,
            c("metformin.rosiglitazone", "time_in_hospital", "readmitted")],
            ycol = "readmitted",
            positive = "YES", is_dropsame = TRUE)
get_woe_plot(res_woe_temp$woe_bins, xcol = c("time_in_hospital"), ncol = 2)
}
