% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd9_map}
\alias{icd9_map}
\title{ICD-9 codes mapping information}
\format{
a dataset with 17 obs. and 3 variables:
\describe{
\item{description}{primary category;}
\item{code_start}{the smallest code for this category;}
\item{code_end}{the largetest code for this category;}
}
}
\source{
\url{https://simba.isr.umich.edu/restricted/docs/Mortality/icd_09_codes.pdf}
}
\description{
the map information for ICD-9 coding and primary category
}
\keyword{data}
