% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IDs_map}
\alias{IDs_map}
\title{ID mapping information for Diabetic Readmission dataset}
\format{
a dataset with 63 obs. and 3 variables:
\describe{
\item{column}{column name;}
\item{value}{value in original dataset, the map_id;}
\item{description}{meaning of map_id;}
}
}
\source{
Beata Strack, Jonathan P. DeShazo, Chris Gennings, Juan L. Olmo,
Sebastian Ventura, Krzysztof J. Cios, and John N. Clore, “Impact of HbA1c
Measurement on Hospital Readmission Rates: Analysis of 70,000 Clinical Database
Patient Records,” BioMed Research International, vol. 2014, Article ID 781670,
11 pages, 2014.
}
\description{
the map information for dataset `diabetic_data`
}
\keyword{data}
