% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woe.R
\name{iv_filter}
\alias{iv_filter}
\title{model selection with iv (Information Value)}
\usage{
iv_filter(data.woe, iv_limit = 0.02, ycol = "readmitted", positive = "YES")
}
\arguments{
\item{data.woe}{dataset, need to be data.frame()}

\item{iv_limit}{delete columns with iv value < iv_limit}

\item{ycol}{the name for the target column}

\item{positive}{the level in target y which denotes the positive class}
}
\value{
`out`, a list with data.iv = processed dataset, iv.info = a data.frame
to show the iv value for each column
}
\description{
This function would do model selection based on iv (Information Value)
for dataset after WOE transformation
}
\examples{
data(diabetic_data)
diabetic_data <- diabetic_data[1:100, c("num_procedures", "readmitted")]
diabetic_data.woe <- build_woe(diabetic_data, ycol = "readmitted",
                               positive = "YES", is.dropsame = TRUE)
iv_res <- iv_filter(diabetic_data.woe, iv_limit = .02, ycol = "readmitted",
                    positive = "YES")
diabetic_data.iv <- iv_res$data.iv
iv.info <- iv_res$iv.info
}
