% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{plot outliers}
\usage{
plot_boxplot(x, ncol = 2)
}
\arguments{
\item{x}{the dataset for plot}

\item{ncol}{the number of plots to show per row}
}
\description{
This function could be used for giving a series of boxplot to
show the possible outliers
}
\examples{
data(diabetic_data)
plot_boxplot(diabetic_data[1:100, c("num_procedures", "num_medications")])
}
