% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_comparehist}
\alias{plot_comparehist}
\title{give histogram plot for given columns}
\usage{
plot_comparehist(data, ycol, xcol, ncol = 4, bins = 10)
}
\arguments{
\item{data}{the dataset for plot}

\item{ycol}{the column name for the target variable}

\item{xcol}{a vector of column names, the function would draw histograms
for each variable in xcol with ycol (the target one)}

\item{ncol}{the number of plots to show per row}

\item{bins}{used for histograms}
}
\description{
This function could be used for giving a series of histograms
for given columns, to explore the relationship between the given variables
and the target column (denote as ycol)
}
\examples{
data(diabetic_data)
plot_comparehist(diabetic_data[1:100, ], "readmitted",
                           c("num_medications"), ncol = 4)
}
