% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{replace_missing_col}
\alias{replace_missing_col}
\title{replacing missing mark in one column with NA}
\usage{
replace_missing_col(
  x,
  mark_list = c("?", "unknown", "invalid", "not available", "not mapped", "null", "none")
)
}
\arguments{
\item{x}{the column needed to do the replacement, if is.numeric(), the replacement
process would be skipped.}

\item{mark_list}{the vector including marks which denotes missing}
}
\value{
the processed column
}
\description{
This function could be use for replacing the missing mark (e.g.
Unknown / invalid / null) with NA for a single column
}
\examples{
data(diabetic_data)
diabetic_data$race <- replace_missing_col(diabetic_data$race, mark_list = c('?'))
}
