% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{replace_missing_dataset}
\alias{replace_missing_dataset}
\title{replacing missing mark in dataset with NA}
\usage{
replace_missing_dataset(
  data,
  mark_list = c("?", "unknown", "invalid", "not available", "not mapped", "null", "none")
)
}
\arguments{
\item{data}{the dataset needed to do the replacement}

\item{mark_list}{the vector including marks which denotes missing}
}
\value{
the processed column
}
\description{
This function could be use for replacing the missing mark (e.g.
Unknown / invalid / null) with NA for a dataset
}
\examples{
data(diabetic_data)
diabetic_data <- replace_missing_dataset(diabetic_data[1:100,],
                                         mark_list = c("?", "null"))
}
