% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsample_smote.R
\name{smote.exs}
\alias{smote.exs}
\title{Obtain a set of smoted examples for a set of rare cases.}
\source{
modified based on L. Torgo, Feb 2010, from `DMwR` package
}
\usage{
smote.exs(data, ycol, perc.over, k)
}
\arguments{
\item{data}{the original training set (rare cases only)}

\item{ycol}{the colname for the target column}

\item{perc.over}{is the 100 * number of new cases (smoted cases) generated for each
rare case. If perc.over < 100 a single case is generated uniquely for a randomly
selected perc.over of the rare cases}

\item{k}{is the number of neighbours to consider as the pool from where
the new examples are generated}
}
\value{
processed dataset (data.frame())
}
\description{
This function would do SMOTE upsampling for a set of rare cases.
}
\examples{
data <- data.frame(x1=rnorm(100),
                   x2=rnorm(100),
                   y=rep(as.factor(c('Yes', 'No')), times=c(90, 10)))
newcase <- smote.exs(data[91:100, ], "y", perc.over = 500, k = 5)
}
