% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsample_smote.R
\name{SMOTE}
\alias{SMOTE}
\title{Creating a SMOTE training sample for classification problems}
\source{
modified based on L. Torgo, Feb 2010, from `DMwR` package
}
\usage{
SMOTE(data, ycol, perc.over = 500, k = 5, perc.under = 200)
}
\arguments{
\item{data}{the original training set (with the unbalanced distribution)}

\item{ycol}{the colname for the target column}

\item{perc.over}{is the 100 * number of new cases (smoted cases) generated for each
rare case. If perc.over < 100 a single case is generated uniquely for a randomly
selected perc.over of the rare cases}

\item{k}{is the number of neighbours to consider as the pool from where
the new examples are generated}

\item{perc.under}{is the number of "normal" cases that are randomly selected for
each smoted case}
}
\value{
processed dataset (data.frame())
}
\description{
This function would do SMOTE upsampling for given dataset with given
sample percent to solve the imbalance problem
}
\examples{
data <- data.frame(y=rep(as.factor(c('Yes', 'No')), times=c(90, 10)),
                 x1=rnorm(100),
                 x2=rnorm(100))
newdata <- SMOTE(data, "y", perc.over = 500, k = 5, perc.under = 200)
}
