% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{ycol_mapped01}
\alias{ycol_mapped01}
\title{mapped target column to 0/1}
\usage{
ycol_mapped01(data, ycol, positive)
}
\arguments{
\item{data}{the dataset needed to do the transformation}

\item{ycol}{String, the colname for the target column}

\item{positive}{String, denotes the positive class}
}
\value{
the processed dataset
}
\description{
This function could be use for mapping the target column for given
dataset to 0/1, as 1 refer to the positive class
}
\examples{
data(diabetic_data)
diabetic_data <- ycol_mapped01(diabetic_data[1:100, ], "readmitted", "YES")
}
