% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{build_logistic_regression}
\alias{build_logistic_regression}
\title{build the logistic regression model for given dataset}
\usage{
build_logistic_regression(
  data,
  ycol,
  is_kfold = FALSE,
  cv_num = 10,
  seed = NA,
  is_smote = FALSE,
  auto_thre = TRUE,
  thre = NA
)
}
\arguments{
\item{data}{dataset for modeling}

\item{ycol}{the colname for target column}

\item{is_kfold}{if TRUE, evaluate the model with k-fold evaluation}

\item{cv_num}{for k = cv_num fold validation}

\item{seed}{use for spliting train/test data}

\item{is_smote}{whether upsampling with smote method for solving imbalance}

\item{auto_thre}{whether learn the best thre automatically based on ROC}

\item{thre}{classification threshold, when the model gives probability
higher than this value, assign it to the positive, otherwise to negative. If
it's NA, then use the information from ROC (specificities & sensitivities)
to calculated automatically.}
}
\value{
`result` a list with `result$mod` trained model; `result$eval_tab`
the model evaluation result; `result$roc_res` the ROC and AUC information;
`result$best_thre` the best threshold used in modeling
}
\description{
This function could be use for training a logistic regression
model for given dataset, with k-fold validation (if applicable) and evaluate
the result
}
\examples{
data(diabetic_data)
diabetic_data <- diabetic_data[1:100, c('time_in_hospital', 'readmitted')] |>
                 ycol_mapped01(ycol = "readmitted", positive = "YES")
res <- build_logistic_regression(diabetic_data,
                         "readmitted", is_kfold = TRUE,
                         cv_num = 2, seed = 103221, is_smote = FALSE,
                         thre = 0.4)
}
