% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woe.R
\name{iv_filter}
\alias{iv_filter}
\title{model selection with iv (Information Value)}
\usage{
iv_filter(data_woe, iv_limit = 0.02, ycol, positive = 1)
}
\arguments{
\item{data_woe}{dataset, need to be data.frame()}

\item{iv_limit}{delete columns with iv value < iv_limit}

\item{ycol}{the name for the target column}

\item{positive}{the level in target y which denotes the positive class}
}
\value{
`out`, a list with data_iv = processed dataset, iv_info =
a data.frame to show the iv value for each column
}
\description{
This function would do model selection based on iv
(Information Value) for dataset after WOE transformation
}
\examples{
data(diabetic_data)
diabetic_data <- diabetic_data[1:100, c("num_procedures", "num_medications",
                                        "readmitted")]
diabetic_data_woe <- build_woe(diabetic_data, ycol = "readmitted",
                               positive = "YES", is_dropsame = TRUE)
iv_res <- iv_filter(diabetic_data_woe$newdata,
                    iv_limit = .01, ycol = "readmitted",
                    positive = "YES")
diabetic_data_iv <- iv_res$data_iv
iv_info <- iv_res$iv_info
}
