% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsample_smote.R
\name{upsample_smote}
\alias{upsample_smote}
\title{upsampling with smote for given dataset}
\usage{
upsample_smote(data_forsample, ycol, perc_over = NA)
}
\arguments{
\item{data_forsample}{dataset need to do upsampling, need to be data.frame}

\item{ycol}{the colname for the target column}

\item{perc_over}{the upsampling percent for SMOTE, is the 100 * number of
new cases (smoted cases) generated for each rare case. If perc_over < 100 a
single case is generated uniquely for a randomly selected perc_over of the
rare cases}
}
\value{
processed dataset (data.frame())
}
\description{
this function would do smote upsampling for given dataset while
calculating appropriate over percentage automatically
}
\examples{
data <- data.frame(y=rep(as.factor(c('Yes', 'No')), times=c(90, 10)),
                 x1=rnorm(100),
                 x2=rnorm(100))
data_smote <- upsample_smote(data, "y")
}
