% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsample_smote.R
\name{upsample_smote}
\alias{upsample_smote}
\title{upsampling with SMOTE for given dataset}
\usage{
upsample_smote(data.forsample, ycol)
}
\arguments{
\item{data.forsample}{dataset need to do upsampling, need to be data.frame}

\item{ycol}{the colname for the target column}
}
\value{
processed dataset (data.frame())
}
\description{
This function would do SMOTE upsampling for given dataset while
calculating appropriate over percentage automatically
}
\examples{
data <- data.frame(y=rep(as.factor(c('Yes', 'No')), times=c(90, 10)),
                 x1=rnorm(100),
                 x2=rnorm(100))
data.smote <- upsample_smote(data, "y")
}
